koan/requirements.txt
requests>=2.28
flask>=3.0
pyyaml>=6.0
ruamel.yaml>=0.18
pytest-timeout>=2.1

# Optional: Slack provider (install with: pip install slack-sdk)
# slack-sdk>=3.27
